requirements.txt
absl-py==0.6.1
astor==0.7.1
bleach==1.5.0
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
Django==2.2.2
djangorestframework==3.9.4
gast==0.2.0
grpcio==1.21.1
h5py==2.9.0
html5lib==0.9999999
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mock==3.0.5
networkx==2.2
numpy==1.15.4
opencv-python==3.4.5.20
Pillow==5.4.0
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2019.1
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1
